Move ConfirmedCasesAttributesCard to components/index/cards
