+++ components/index/cards/ConfirmedCasesAttributesCard.vue
